feat: add defaults for donation
